Moved front behaviour into it's own package.
It is one of our more complex pieces of code and I want it firewalled off.
