+++ front/app/controllers/front/TrailblockAgent.scala
@@ -1,12 +1,12 @@
-package controllers
+package controllers.front
 
 import model._
 import common.{ Logging, AkkaSupport }
+import conf.ContentApi
+import com.gu.openplatform.contentapi.model.ItemResponse
 import model.Trailblock
 import scala.Some
 import model.TrailblockDescription
-import com.gu.openplatform.contentapi.model.ItemResponse
-import conf.ContentApi
 
 /*
   Responsible for refreshing one block on the front (e.g. the Sport block) for one edition
@@ -15,7 +15,9 @@ class TrailblockAgent(val description: TrailblockDescription, edition: String) e
 
   private lazy val agent = play_akka.agent[Option[Trailblock]](None)
 
-  def refresh() = agent.sendOff { old => Some(Trailblock(description, loadTrails(description.id))) }
+  def refresh() = agent.sendOff {
+    old => Some(Trailblock(description, loadTrails(description.id)))
+  }
 
   def close() = agent.close()
 
@@ -29,9 +31,13 @@ class TrailblockAgent(val description: TrailblockDescription, edition: String) e
       .pageSize(20)
       .response
 
-    val editorsPicks = response.editorsPicks map { new Content(_) }
+    val editorsPicks = response.editorsPicks map {
+      new Content(_)
+    }
     val editorsPicksIds = editorsPicks map (_.id)
-    val latest = response.results map { new Content(_) } filterNot (c => editorsPicksIds contains (c.id))
+    val latest = response.results map {
+      new Content(_)
+    } filterNot (c => editorsPicksIds contains (c.id))
 
     editorsPicks ++ latest
   }
